test: split remaining install and prepare runner coverage
Break the last oversized install and prepare runner test files into concern-focused units so runtime behavior stays covered while review and navigation costs keep dropping.
