+++ internal/install/output_contracts_test.go
@@ -97,12 +97,3 @@ func writePrivateTestFile(path string, content []byte) error {
 func coverageKey(kind, output string) string {
 	return kind + ":" + output
 }
-
-func contains(values []string, want string) bool {
-	for _, value := range values {
-		if value == want {
-			return true
-		}
-	}
-	return false
-}
