+++ internal/install/run_lifecycle_test.go
@@ -1,0 +1,409 @@
+package install
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+
+	"github.com/Airgap-Castaways/deck/internal/config"
+	"github.com/Airgap-Castaways/deck/internal/errcode"
+)
+
+func TestRun_InstallTools(t *testing.T) {
+	dir := t.TempDir()
+	statePath := filepath.Join(dir, "state", "state.json")
+	bundle := filepath.Join(dir, "bundle")
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("ok")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+
+	fileA := filepath.Join(dir, "a.txt")
+	fileB := filepath.Join(dir, "b.txt")
+	sysctlPath := filepath.Join(dir, "sysctl.conf")
+	modprobePath := filepath.Join(dir, "modules.conf")
+	joinPath := filepath.Join(dir, "join.txt")
+	binDir := filepath.Join(dir, "bin")
+	if err := os.MkdirAll(binDir, 0o755); err != nil {
+		t.Fatalf("mkdir bin: %v", err)
+	}
+	fakeApt := filepath.Join(binDir, "apt-get")
+	if err := os.WriteFile(fakeApt, []byte("#!/usr/bin/env bash\nset -euo pipefail\nexit 0\n"), 0o755); err != nil {
+		t.Fatalf("write fake apt-get: %v", err)
+	}
+	originalPath := os.Getenv("PATH")
+	t.Setenv("PATH", fmt.Sprintf("%s:%s", binDir, originalPath))
+	kubeadm := useStubInitJoinKubeadm()
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "install",
+			Steps: []config.Step{
+				{ID: "install-packages", Kind: "InstallPackage", Spec: map[string]any{"packages": []any{"containerd"}}},
+				{ID: "write-file", Kind: "WriteFile", Spec: map[string]any{"path": fileA, "content": "hello world"}},
+				{ID: "edit-file", Kind: "EditFile", Spec: map[string]any{"path": fileA, "edits": []any{map[string]any{"op": "replace", "match": "world", "replaceWith": "deck"}}}},
+				{ID: "copy-file", Kind: "CopyFile", Spec: map[string]any{"source": map[string]any{"path": fileA}, "path": fileB}},
+				{ID: "Sysctl", Kind: "Sysctl", Spec: map[string]any{"writeFile": sysctlPath, "values": map[string]any{"net.ipv4.ip_forward": "1"}}},
+				{ID: "modprobe", Kind: "KernelModule", Spec: map[string]any{"name": "overlay", "persistFile": modprobePath}},
+				{ID: "run-cmd", Kind: "Command", Spec: map[string]any{"command": []any{"true"}}},
+				{ID: "kubeadm-init", Kind: "InitKubeadm", Spec: map[string]any{"outputJoinFile": joinPath}},
+				{ID: "kubeadm-join", Kind: "JoinKubeadm", Spec: map[string]any{"joinFile": joinPath}},
+			},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath, kubeadm: kubeadm}); err != nil {
+		t.Fatalf("Run failed: %v", err)
+	}
+
+	contentA, err := os.ReadFile(fileA)
+	if err != nil {
+		t.Fatalf("read fileA: %v", err)
+	}
+	if string(contentA) != "hello deck\n" {
+		t.Fatalf("unexpected edited content: %q", string(contentA))
+	}
+
+	if _, err := os.Stat(fileB); err != nil {
+		t.Fatalf("copied file missing: %v", err)
+	}
+	if _, err := os.Stat(sysctlPath); err != nil {
+		t.Fatalf("sysctl file missing: %v", err)
+	}
+	if _, err := os.Stat(modprobePath); err != nil {
+		t.Fatalf("modprobe persist file missing: %v", err)
+	}
+	if _, err := os.Stat(joinPath); err != nil {
+		t.Fatalf("join file missing: %v", err)
+	}
+
+	rawState, err := os.ReadFile(statePath)
+	if err != nil {
+		t.Fatalf("read state: %v", err)
+	}
+	var st State
+	if err := json.Unmarshal(rawState, &st); err != nil {
+		t.Fatalf("parse state: %v", err)
+	}
+	if len(st.CompletedPhases) != 1 || st.CompletedPhases[0] != "install" {
+		t.Fatalf("unexpected completed phases: %#v", st.CompletedPhases)
+	}
+	if st.Phase != "completed" {
+		t.Fatalf("expected final phase state completed, got %q", st.Phase)
+	}
+}
+
+func TestRun_DefaultStatePathUsesHomeStateKey(t *testing.T) {
+	dir := t.TempDir()
+	home := filepath.Join(dir, "home")
+	bundle := filepath.Join(dir, "bundle")
+	if err := os.MkdirAll(home, 0o755); err != nil {
+		t.Fatalf("mkdir home: %v", err)
+	}
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("ok")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+	t.Setenv("HOME", home)
+
+	wf := &config.Workflow{
+		StateKey: "state-key-default-path-test",
+		Phases: []config.Phase{{
+			Name:  "install",
+			Steps: []config.Step{{ID: "s1", Kind: "Command", Spec: map[string]any{"command": []any{"true"}}}},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle}); err != nil {
+		t.Fatalf("Run failed: %v", err)
+	}
+
+	statePath := filepath.Join(home, ".local", "state", "deck", "state", wf.StateKey+".json")
+	if _, err := os.Stat(statePath); err != nil {
+		t.Fatalf("state file missing at expected home path: %v", err)
+	}
+	if _, err := os.Stat(filepath.Join(bundle, ".deck", "state.json")); !os.IsNotExist(err) {
+		t.Fatalf("unexpected bundle state file, err=%v", err)
+	}
+}
+
+func TestRun_ManifestIntegrityVerified(t *testing.T) {
+	dir := t.TempDir()
+	bundle := filepath.Join(dir, "bundle")
+	statePath := filepath.Join(dir, "state", "state.json")
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("hello"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("hello")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name:  "install",
+			Steps: []config.Step{{ID: "s1", Kind: "Command", Spec: map[string]any{"command": []any{"true"}}}},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath}); err != nil {
+		t.Fatalf("expected install success, got %v", err)
+	}
+}
+
+func TestRun_ManifestIntegrityMismatch(t *testing.T) {
+	dir := t.TempDir()
+	bundle := filepath.Join(dir, "bundle")
+	statePath := filepath.Join(dir, "state", "state.json")
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("hello"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("different")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name:  "install",
+			Steps: []config.Step{{ID: "s1", Kind: "Command", Spec: map[string]any{"command": []any{"true"}}}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath})
+	if err == nil {
+		t.Fatalf("expected manifest integrity error")
+	}
+	if !strings.Contains(err.Error(), "E_BUNDLE_INTEGRITY") {
+		t.Fatalf("expected E_BUNDLE_INTEGRITY error, got %v", err)
+	}
+}
+
+func TestRun_NoPhasesFails(t *testing.T) {
+	err := Run(context.Background(), &config.Workflow{Version: "v1"}, RunOptions{})
+	if err == nil {
+		t.Fatalf("expected no phases error")
+	}
+	if !strings.Contains(err.Error(), "no phases found") {
+		t.Fatalf("unexpected error: %v", err)
+	}
+}
+
+func TestRun_ResumeFromFailedStep(t *testing.T) {
+	dir := t.TempDir()
+	statePath := filepath.Join(dir, "state", "state.json")
+	bundle := filepath.Join(dir, "bundle")
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("ok")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+
+	first := filepath.Join(dir, "first.txt")
+	second := filepath.Join(dir, "second.txt")
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "install",
+			Steps: []config.Step{
+				{ID: "s1", Kind: "WriteFile", Spec: map[string]any{"path": first, "content": "ok"}},
+				{ID: "s2", Kind: "Command", Spec: map[string]any{"command": []any{"false"}}},
+				{ID: "s3", Kind: "WriteFile", Spec: map[string]any{"path": second, "content": "done"}},
+			},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath}); err == nil {
+		t.Fatalf("expected failure on s2")
+	}
+
+	rawState, err := os.ReadFile(statePath)
+	if err != nil {
+		t.Fatalf("read state: %v", err)
+	}
+	var st State
+	if err := json.Unmarshal(rawState, &st); err != nil {
+		t.Fatalf("parse state: %v", err)
+	}
+	if len(st.CompletedPhases) != 0 {
+		t.Fatalf("unexpected completed phases after failure: %#v", st.CompletedPhases)
+	}
+	if st.FailedPhase != "install" {
+		t.Fatalf("expected failed phase install, got %q", st.FailedPhase)
+	}
+
+	wf.Phases[0].Steps[1].Spec = map[string]any{"command": []any{"true"}}
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath}); err != nil {
+		t.Fatalf("resume run failed: %v", err)
+	}
+
+	if _, err := os.Stat(second); err != nil {
+		t.Fatalf("expected second file after resume: %v", err)
+	}
+
+	rawState, err = os.ReadFile(statePath)
+	if err != nil {
+		t.Fatalf("read final state: %v", err)
+	}
+	var final State
+	if err := json.Unmarshal(rawState, &final); err != nil {
+		t.Fatalf("parse final state: %v", err)
+	}
+	if len(final.CompletedPhases) != 1 || final.CompletedPhases[0] != "install" {
+		t.Fatalf("unexpected completed phases after resume: %#v", final.CompletedPhases)
+	}
+	if final.FailedPhase != "" {
+		t.Fatalf("expected empty failed phase, got %q", final.FailedPhase)
+	}
+}
+
+func TestRun_UnsupportedInstallKindFails(t *testing.T) {
+	dir := t.TempDir()
+	bundle := filepath.Join(dir, "bundle")
+	statePath := filepath.Join(dir, "state", "state.json")
+	if err := os.MkdirAll(bundle, 0o755); err != nil {
+		t.Fatalf("mkdir bundle: %v", err)
+	}
+	artifact := filepath.Join(bundle, "files", "a.txt")
+	if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+		t.Fatalf("mkdir files: %v", err)
+	}
+	if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+		t.Fatalf("write artifact: %v", err)
+	}
+	if err := writeManifestForTest(bundle, "files/a.txt", []byte("ok")); err != nil {
+		t.Fatalf("write manifest: %v", err)
+	}
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name:  "install",
+			Steps: []config.Step{{ID: "x", Kind: "UnknownKind", Spec: map[string]any{}}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath})
+	if err == nil {
+		t.Fatalf("expected unsupported kind error")
+	}
+	if !strings.Contains(err.Error(), "E_INSTALL_KIND_UNSUPPORTED") {
+		t.Fatalf("expected E_INSTALL_KIND_UNSUPPORTED, got %v", err)
+	}
+}
+
+func TestRun_ExposesTypedErrorCodes(t *testing.T) {
+	t.Run("unsupported kind", func(t *testing.T) {
+		dir := t.TempDir()
+		bundle := filepath.Join(dir, "bundle")
+		statePath := filepath.Join(dir, "state", "state.json")
+		if err := os.MkdirAll(bundle, 0o755); err != nil {
+			t.Fatalf("mkdir bundle: %v", err)
+		}
+		artifact := filepath.Join(bundle, "files", "placeholder.txt")
+		if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+			t.Fatalf("mkdir files: %v", err)
+		}
+		if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+			t.Fatalf("write artifact: %v", err)
+		}
+		if err := writeManifestForTest(bundle, "files/placeholder.txt", []byte("ok")); err != nil {
+			t.Fatalf("write manifest: %v", err)
+		}
+
+		wf := &config.Workflow{
+			Version: "v1",
+			Phases: []config.Phase{{
+				Name:  "install",
+				Steps: []config.Step{{ID: "x", Kind: "UnknownKind", Spec: map[string]any{}}},
+			}},
+		}
+
+		err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath})
+		if !errcode.Is(err, errCodeInstallKindUnsupported) {
+			t.Fatalf("expected typed code %s, got %v", errCodeInstallKindUnsupported, err)
+		}
+	})
+
+	t.Run("command timeout", func(t *testing.T) {
+		dir := t.TempDir()
+		bundle := filepath.Join(dir, "bundle")
+		statePath := filepath.Join(dir, "state", "state.json")
+		if err := os.MkdirAll(bundle, 0o755); err != nil {
+			t.Fatalf("mkdir bundle: %v", err)
+		}
+		artifact := filepath.Join(bundle, "files", "a.txt")
+		if err := os.MkdirAll(filepath.Dir(artifact), 0o755); err != nil {
+			t.Fatalf("mkdir files: %v", err)
+		}
+		if err := os.WriteFile(artifact, []byte("ok"), 0o644); err != nil {
+			t.Fatalf("write artifact: %v", err)
+		}
+		if err := writeManifestForTest(bundle, "files/a.txt", []byte("ok")); err != nil {
+			t.Fatalf("write manifest: %v", err)
+		}
+
+		wf := &config.Workflow{
+			Version: "v1",
+			Phases: []config.Phase{{
+				Name: "install",
+				Steps: []config.Step{{
+					ID:   "cmd",
+					Kind: "Command",
+					Spec: map[string]any{"command": []any{"sleep", "1"}, "timeout": "10ms"},
+				}},
+			}},
+		}
+
+		err := Run(context.Background(), wf, RunOptions{BundleRoot: bundle, StatePath: statePath})
+		if !errcode.Is(err, errCodeInstallCommandTimeout) {
+			t.Fatalf("expected typed code %s, got %v", errCodeInstallCommandTimeout, err)
+		}
+	})
+}
