+++ internal/prepare/download_file_fallback_test.go
@@ -1,0 +1,288 @@
+package prepare
+
+import (
+	"context"
+	"crypto/sha256"
+	"encoding/hex"
+	"net/http"
+	"net/http/httptest"
+	"os"
+	"path/filepath"
+	"strings"
+	"testing"
+	"time"
+
+	"github.com/Airgap-Castaways/deck/internal/config"
+	"github.com/Airgap-Castaways/deck/internal/errcode"
+)
+
+func TestRun_FileFallbackLocalThenBundle(t *testing.T) {
+	bundleOut := t.TempDir()
+	localCache := t.TempDir()
+	bundleCache := t.TempDir()
+
+	relSource := filepath.ToSlash(filepath.Join("files", "artifact.bin"))
+	bundleOnlyPath := filepath.Join(bundleCache, filepath.FromSlash(relSource))
+	if err := os.MkdirAll(filepath.Dir(bundleOnlyPath), 0o755); err != nil {
+		t.Fatalf("mkdir bundle cache path: %v", err)
+	}
+	if err := os.WriteFile(bundleOnlyPath, []byte("from-bundle-source"), 0o644); err != nil {
+		t.Fatalf("write bundle cache source: %v", err)
+	}
+	sum := sha256.Sum256([]byte("from-bundle-source"))
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source": map[string]any{
+						"path":   relSource,
+						"sha256": hex.EncodeToString(sum[:]),
+					},
+					"fetch": map[string]any{
+						"strategy": "fallback",
+						"sources": []any{
+							map[string]any{"type": "local", "path": localCache},
+							map[string]any{"type": "bundle", "path": bundleCache},
+						},
+					},
+					"outputPath": "files/fetched.bin",
+				},
+			}},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut}); err != nil {
+		t.Fatalf("Run failed: %v", err)
+	}
+
+	raw, err := os.ReadFile(filepath.Join(bundleOut, "files", "fetched.bin"))
+	if err != nil {
+		t.Fatalf("read fetched output: %v", err)
+	}
+	if string(raw) != "from-bundle-source" {
+		t.Fatalf("unexpected fetched content: %q", string(raw))
+	}
+}
+
+func TestRun_FileFallbackSourceMissing(t *testing.T) {
+	bundleOut := t.TempDir()
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source": map[string]any{
+						"path": "files/missing.bin",
+					},
+					"fetch": map[string]any{
+						"strategy": "fallback",
+						"sources":  []any{map[string]any{"type": "local", "path": t.TempDir()}},
+					},
+					"outputPath": "files/out.bin",
+				},
+			}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut})
+	if err == nil {
+		t.Fatalf("expected source not found error")
+	}
+	if !errcode.Is(err, errCodeArtifactSourceNotFound) {
+		t.Fatalf("expected typed code %s, got %v", errCodeArtifactSourceNotFound, err)
+	}
+	if !strings.Contains(err.Error(), "E_PREPARE_SOURCE_NOT_FOUND") {
+		t.Fatalf("expected E_PREPARE_SOURCE_NOT_FOUND, got %v", err)
+	}
+}
+
+func TestRun_DownloadFileRejectsNonCanonicalOutputPath(t *testing.T) {
+	bundleOut := t.TempDir()
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source":     map[string]any{"url": "https://example.invalid/artifact.bin"},
+					"outputPath": "artifacts/out.bin",
+				},
+			}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut})
+	if err == nil || !strings.Contains(err.Error(), "DownloadFile outputPath must stay under files/") {
+		t.Fatalf("expected canonical outputPath error, got %v", err)
+	}
+}
+
+func TestResolveSourceBytes_PreservesContextCancellation(t *testing.T) {
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		time.Sleep(250 * time.Millisecond)
+		_, _ = w.Write([]byte("should-not-complete"))
+	}))
+	defer server.Close()
+
+	ctx, cancel := context.WithCancel(context.Background())
+	cancel()
+
+	_, err := resolveSourceBytes(ctx, map[string]any{
+		"fetch": map[string]any{
+			"sources": []any{map[string]any{"type": "online", "url": server.URL}},
+		},
+	}, "files/remote.bin")
+	if err == nil {
+		t.Fatalf("expected context cancellation error")
+	}
+	if strings.Contains(err.Error(), "E_PREPARE_SOURCE_NOT_FOUND") {
+		t.Fatalf("expected cancellation to not be mapped to source-not-found, got %v", err)
+	}
+	if !strings.Contains(err.Error(), context.Canceled.Error()) {
+		t.Fatalf("expected canceled context in error, got %v", err)
+	}
+}
+
+func TestRun_FileFallbackRepoThenOnline(t *testing.T) {
+	bundleOut := t.TempDir()
+
+	repo := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		http.NotFound(w, r)
+	}))
+	defer repo.Close()
+
+	online := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		if r.URL.Path == "/files/remote.bin" {
+			_, _ = w.Write([]byte("from-online-source"))
+			return
+		}
+		http.NotFound(w, r)
+	}))
+	defer online.Close()
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source": map[string]any{
+						"path": "files/remote.bin",
+					},
+					"fetch": map[string]any{
+						"strategy": "fallback",
+						"sources": []any{
+							map[string]any{"type": "repo", "url": repo.URL},
+							map[string]any{"type": "online", "url": online.URL},
+						},
+					},
+					"outputPath": "files/fetched-online.bin",
+				},
+			}},
+		}},
+	}
+
+	if err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut}); err != nil {
+		t.Fatalf("Run failed: %v", err)
+	}
+
+	raw, err := os.ReadFile(filepath.Join(bundleOut, "files", "fetched-online.bin"))
+	if err != nil {
+		t.Fatalf("read fetched output: %v", err)
+	}
+	if string(raw) != "from-online-source" {
+		t.Fatalf("unexpected fetched content: %q", string(raw))
+	}
+}
+
+func TestRun_FileOfflinePolicyBlocksOnlineFallback(t *testing.T) {
+	bundleOut := t.TempDir()
+
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		_, _ = w.Write([]byte("should-not-be-downloaded"))
+	}))
+	defer server.Close()
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source": map[string]any{
+						"path": "files/not-found.bin",
+						"url":  server.URL + "/files/not-found.bin",
+					},
+					"fetch": map[string]any{
+						"offlineOnly": true,
+						"strategy":    "fallback",
+						"sources":     []any{map[string]any{"type": "online", "url": server.URL}},
+					},
+					"outputPath": "files/out.bin",
+				},
+			}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut})
+	if err == nil {
+		t.Fatalf("expected offline policy block error")
+	}
+	if !errcode.Is(err, errCodePrepareOfflinePolicyBlock) {
+		t.Fatalf("expected typed code %s, got %v", errCodePrepareOfflinePolicyBlock, err)
+	}
+	if !strings.Contains(err.Error(), "E_PREPARE_OFFLINE_POLICY_BLOCK") {
+		t.Fatalf("expected E_PREPARE_OFFLINE_POLICY_BLOCK, got %v", err)
+	}
+}
+
+func TestRun_FileOfflinePolicyBlocksDirectURL(t *testing.T) {
+	bundleOut := t.TempDir()
+
+	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		_, _ = w.Write([]byte("should-not-be-downloaded"))
+	}))
+	defer server.Close()
+
+	wf := &config.Workflow{
+		Version: "v1",
+		Phases: []config.Phase{{
+			Name: "prepare",
+			Steps: []config.Step{{
+				ID:   "download-file",
+				Kind: "DownloadFile",
+				Spec: map[string]any{
+					"source":     map[string]any{"url": server.URL + "/files/a.bin"},
+					"fetch":      map[string]any{"offlineOnly": true},
+					"outputPath": "files/out.bin",
+				},
+			}},
+		}},
+	}
+
+	err := Run(context.Background(), wf, RunOptions{BundleRoot: bundleOut})
+	if err == nil {
+		t.Fatalf("expected offline policy block error")
+	}
+	if !errcode.Is(err, errCodePrepareOfflinePolicyBlock) {
+		t.Fatalf("expected typed code %s, got %v", errCodePrepareOfflinePolicyBlock, err)
+	}
+	if !strings.Contains(err.Error(), "E_PREPARE_OFFLINE_POLICY_BLOCK") {
+		t.Fatalf("expected E_PREPARE_OFFLINE_POLICY_BLOCK, got %v", err)
+	}
+}
